- fixes after review: split large files contents, renaming, optimization in staking process
