+++ vm/systemSmartContracts/common.go
@@ -1,6 +1,12 @@
 package systemSmartContracts
 
 import (
+	"bytes"
+	"encoding/hex"
+	"fmt"
+	"math/big"
+	"math/rand"
+
 	"github.com/ElrondNetwork/elrond-go/vm"
 	vmcommon "github.com/ElrondNetwork/elrond-vm-common"
 )
@@ -24,4 +30,101 @@ func CheckIfNil(args *vmcommon.ContractCallInput) error {
 	}
 
 	return nil
+}
+
+func verifyBLSPublicKeys(registrationData *AuctionDataV2, arguments [][]byte) error {
+	for _, argKey := range arguments {
+		found := false
+		for _, blsKey := range registrationData.BlsPubKeys {
+			if bytes.Equal(argKey, blsKey) {
+				found = true
+				break
+			}
+		}
+
+		if !found {
+			return fmt.Errorf("%w, key %s not found", vm.ErrBLSPublicKeyMismatch, hex.EncodeToString(argKey))
+		}
+	}
+
+	return nil
+}
+
+func calcTotalQualifyingStake(nodePrice *big.Int, bids []AuctionDataV2) *big.Int {
+	totalQualifyingStake := big.NewInt(0)
+	for _, validator := range bids {
+		if validator.MaxStakePerNode.Cmp(nodePrice) < 0 {
+			continue
+		}
+
+		maxPossibleNodes := big.NewInt(0).Div(validator.TotalStakeValue, nodePrice)
+		if maxPossibleNodes.Uint64() > uint64(len(validator.BlsPubKeys)) {
+			validatorQualifyingStake := big.NewInt(0).Mul(nodePrice, big.NewInt(int64(len(validator.BlsPubKeys))))
+			totalQualifyingStake.Add(totalQualifyingStake, validatorQualifyingStake)
+		} else {
+			totalQualifyingStake.Add(totalQualifyingStake, validator.TotalStakeValue)
+		}
+	}
+
+	return totalQualifyingStake
+}
+
+func calcNumQualifiedNodes(nodePrice *big.Int, bids []AuctionDataV2) uint32 {
+	numQualifiedNodes := uint32(0)
+	for _, validator := range bids {
+		if validator.MaxStakePerNode.Cmp(nodePrice) < 0 {
+			continue
+		}
+		if validator.TotalStakeValue.Cmp(nodePrice) < 0 {
+			continue
+		}
+
+		maxPossibleNodes := big.NewInt(0).Div(validator.TotalStakeValue, nodePrice)
+		if maxPossibleNodes.Uint64() > uint64(len(validator.BlsPubKeys)) {
+			numQualifiedNodes += uint32(len(validator.BlsPubKeys))
+		} else {
+			numQualifiedNodes += uint32(maxPossibleNodes.Uint64())
+		}
+	}
+
+	return numQualifiedNodes
+}
+
+func calcNumAllocatedAndProportion(
+	validator AuctionDataV2,
+	nodePrice *big.Int,
+	totalQualifyingStake *big.Float,
+) (uint64, float64) {
+	maxPossibleNodes := big.NewInt(0).Div(validator.TotalStakeValue, nodePrice)
+	validatorQualifyingStake := big.NewFloat(0).SetInt(validator.TotalStakeValue)
+	qualifiedNodes := maxPossibleNodes.Uint64()
+
+	if maxPossibleNodes.Uint64() > uint64(len(validator.BlsPubKeys)) {
+		validatorQualifyingStake = big.NewFloat(0).SetInt(big.NewInt(0).Mul(nodePrice, big.NewInt(int64(len(validator.BlsPubKeys)))))
+		qualifiedNodes = uint64(len(validator.BlsPubKeys))
+	}
+
+	proportionOfTotalStake := big.NewFloat(0).Quo(validatorQualifyingStake, totalQualifyingStake)
+	proportion, _ := proportionOfTotalStake.Float64()
+	allocatedNodes := float64(qualifiedNodes) * proportion
+	numAllocatedNodes := uint64(allocatedNodes)
+
+	return numAllocatedNodes, allocatedNodes
+}
+
+func shuffleList(list []string, random []byte) {
+	randomSeed := big.NewInt(0).SetBytes(random[:8])
+	r := rand.New(rand.NewSource(randomSeed.Int64()))
+
+	for n := len(list); n > 0; n-- {
+		randIndex := r.Intn(n)
+		list[n-1], list[randIndex] = list[randIndex], list[n-1]
+	}
+}
+
+func isNumArgsCorrectToStake(args [][]byte) bool {
+	maxNodesToRun := big.NewInt(0).SetBytes(args[0]).Uint64()
+	areEnoughArgs := uint64(len(args)) >= 2*maxNodesToRun+1       // NumNodes + LIST(BLS_KEY+SignedMessage)
+	areNotTooManyArgs := uint64(len(args)) <= 2*maxNodesToRun+1+2 // +2 are the optionals - reward address, maxStakePerNode
+	return areEnoughArgs && areNotTooManyArgs
 }
