+++ vm/systemSmartContracts/staking.go
@@ -41,9 +41,10 @@ type stakingSC struct {
 	marshalizer              marshal.Marshalizer
 	enableStakingEpoch       uint32
 	stakeValue               *big.Int
-	flagStake                atomic.Flag
-	flagStakingV2            atomic.Flag
+	flagEnableStaking        atomic.Flag
+	flagSetOwner             atomic.Flag
 	stakingV2Epoch           uint32
+	walletAddressLen         int
 }
 
 // ArgsNewStakingSmartContract holds the arguments needed to create a StakingSmartContract
@@ -109,6 +110,7 @@ func NewStakingSmartContract(
 		endOfEpochAccessAddr:     args.EndOfEpochAccessAddr,
 		enableStakingEpoch:       args.StakingSCConfig.StakeEnableEpoch,
 		stakingV2Epoch:           args.StakingSCConfig.StakingV2Epoch,
+		walletAddressLen:         len(args.StakingAccessAddr),
 	}
 
 	conversionOk := true
@@ -174,42 +176,6 @@ func (r *stakingSC) Execute(args *vmcommon.ContractCallInput) vmcommon.ReturnCod
 	}
 
 	return vmcommon.UserError
-}
-
-func (r *stakingSC) getConfig() *StakingNodesConfig {
-	stakeConfig := &StakingNodesConfig{
-		MinNumNodes: int64(r.minNumNodes),
-		MaxNumNodes: int64(r.maxNumNodes),
-	}
-	configData := r.eei.GetStorage([]byte(nodesConfigKey))
-	if len(configData) == 0 {
-		return stakeConfig
-	}
-
-	err := r.marshalizer.Unmarshal(stakeConfig, configData)
-	if err != nil {
-		log.Warn("unmarshal error on getConfig function, returning baseConfig",
-			"error", err.Error(),
-		)
-		return &StakingNodesConfig{
-			MinNumNodes: int64(r.minNumNodes),
-			MaxNumNodes: int64(r.maxNumNodes),
-		}
-	}
-
-	return stakeConfig
-}
-
-func (r *stakingSC) setConfig(config *StakingNodesConfig) {
-	configData, err := r.marshalizer.Marshal(config)
-	if err != nil {
-		log.Warn("marshal error on setConfig function",
-			"error", err.Error(),
-		)
-		return
-	}
-
-	r.eei.SetStorage([]byte(nodesConfigKey), configData)
 }
 
 func (r *stakingSC) addToStakedNodes() {
@@ -292,7 +258,7 @@ func (r *stakingSC) changeRewardAddress(args *vmcommon.ContractCallInput) vmcomm
 	}
 
 	newRewardAddress := args.Arguments[0]
-	if len(newRewardAddress) != len(args.CallerAddr) {
+	if len(newRewardAddress) != r.walletAddressLen {
 		r.eei.AddReturnMessage("invalid reward address")
 		return vmcommon.UserError
 	}
@@ -361,7 +327,7 @@ func (r *stakingSC) unJailV1(args *vmcommon.ContractCallInput) vmcommon.ReturnCo
 }
 
 func (r *stakingSC) unJail(args *vmcommon.ContractCallInput) vmcommon.ReturnCode {
-	if !r.flagStake.IsSet() {
+	if !r.flagEnableStaking.IsSet() {
 		return r.unJailV1(args)
 	}
 
@@ -404,116 +370,6 @@ func (r *stakingSC) unJail(args *vmcommon.ContractCallInput) vmcommon.ReturnCode
 	}
 
 	return vmcommon.Ok
-}
-
-func (r *stakingSC) getOrCreateRegisteredData(key []byte) (*StakedDataV2_0, error) {
-	registrationData := &StakedDataV2_0{
-		RegisterNonce: 0,
-		Staked:        false,
-		UnStakedNonce: 0,
-		UnStakedEpoch: core.DefaultUnstakedEpoch,
-		RewardAddress: nil,
-		StakeValue:    big.NewInt(0),
-		JailedRound:   math.MaxUint64,
-		UnJailedNonce: 0,
-		JailedNonce:   0,
-		StakedNonce:   math.MaxUint64,
-		SlashValue:    big.NewInt(0),
-	}
-
-	data := r.eei.GetStorage(key)
-	if len(data) > 0 {
-		err := r.marshalizer.Unmarshal(registrationData, data)
-		if err != nil {
-			log.Debug("unmarshal error on staking SC stake function",
-				"error", err.Error(),
-			)
-			return nil, err
-		}
-	}
-	if registrationData.SlashValue == nil {
-		registrationData.SlashValue = big.NewInt(0)
-	}
-
-	return registrationData, nil
-}
-
-func (r *stakingSC) saveStakingData(key []byte, stakedData *StakedDataV2_0) error {
-	if !r.flagStake.IsSet() {
-		return r.saveAsStakingDataV1P0(key, stakedData)
-	}
-	if !r.flagStakingV2.IsSet() {
-		return r.saveAsStakingDataV1P1(key, stakedData)
-	}
-
-	data, err := r.marshalizer.Marshal(stakedData)
-	if err != nil {
-		log.Debug("marshal error in saveStakingData ",
-			"error", err.Error(),
-		)
-		return err
-	}
-
-	r.eei.SetStorage(key, data)
-	return nil
-}
-
-func (r *stakingSC) saveAsStakingDataV1P0(key []byte, stakedData *StakedDataV2_0) error {
-	stakedDataV1 := &StakedDataV1_0{
-		RegisterNonce: stakedData.RegisterNonce,
-		StakedNonce:   stakedData.StakedNonce,
-		Staked:        stakedData.Staked,
-		UnStakedNonce: stakedData.UnStakedNonce,
-		UnStakedEpoch: stakedData.UnStakedEpoch,
-		RewardAddress: stakedData.RewardAddress,
-		StakeValue:    stakedData.StakeValue,
-		JailedRound:   stakedData.JailedRound,
-		JailedNonce:   stakedData.JailedNonce,
-		UnJailedNonce: stakedData.UnJailedNonce,
-		Jailed:        stakedData.Jailed,
-		Waiting:       stakedData.Waiting,
-	}
-
-	data, err := r.marshalizer.Marshal(stakedDataV1)
-	if err != nil {
-		log.Debug("marshal error in saveAsStakingDataV1P0 ",
-			"error", err.Error(),
-		)
-		return err
-	}
-
-	r.eei.SetStorage(key, data)
-	return nil
-}
-
-func (r *stakingSC) saveAsStakingDataV1P1(key []byte, stakedData *StakedDataV2_0) error {
-	stakedDataV1 := &StakedDataV1_1{
-		RegisterNonce: stakedData.RegisterNonce,
-		StakedNonce:   stakedData.StakedNonce,
-		Staked:        stakedData.Staked,
-		UnStakedNonce: stakedData.UnStakedNonce,
-		UnStakedEpoch: stakedData.UnStakedEpoch,
-		RewardAddress: stakedData.RewardAddress,
-		StakeValue:    stakedData.StakeValue,
-		JailedRound:   stakedData.JailedRound,
-		JailedNonce:   stakedData.JailedNonce,
-		UnJailedNonce: stakedData.UnJailedNonce,
-		Jailed:        stakedData.Jailed,
-		Waiting:       stakedData.Waiting,
-		NumJailed:     stakedData.NumJailed,
-		SlashValue:    stakedData.SlashValue,
-	}
-
-	data, err := r.marshalizer.Marshal(stakedDataV1)
-	if err != nil {
-		log.Debug("marshal error in saveAsStakingDataV1P1 ",
-			"error", err.Error(),
-		)
-		return err
-	}
-
-	r.eei.SetStorage(key, data)
-	return nil
 }
 
 func (r *stakingSC) jail(args *vmcommon.ContractCallInput) vmcommon.ReturnCode {
@@ -587,8 +443,8 @@ func (r *stakingSC) stake(args *vmcommon.ContractCallInput, onlyRegister bool) v
 		r.eei.AddReturnMessage("stake function not allowed to be called by address " + string(args.CallerAddr))
 		return vmcommon.UserError
 	}
-	if len(args.Arguments) < 2 {
-		r.eei.AddReturnMessage("not enough arguments, needed BLS key and reward address")
+	if len(args.Arguments) < 3 {
+		r.eei.AddReturnMessage("not enough arguments, needed BLS key, reward address and owner address")
 		return vmcommon.UserError
 	}
 
@@ -603,6 +459,7 @@ func (r *stakingSC) stake(args *vmcommon.ContractCallInput, onlyRegister bool) v
 	}
 
 	registrationData.RewardAddress = args.Arguments[1]
+	registrationData.OwnerAddress = args.Arguments[2]
 	registrationData.StakeValue.Set(r.stakeValue)
 	if !onlyRegister {
 		err = r.processStake(args.Arguments[0], registrationData, false)
@@ -1460,8 +1317,7 @@ func (r *stakingSC) getWaitingListRegisterNonceAndRewardAddress(args *vmcommon.C
 }
 
 func (r *stakingSC) setOwner(args *vmcommon.ContractCallInput) vmcommon.ReturnCode {
-	if !r.flagStakingV2.IsSet() {
-		//setOwner function will become available after enabling staking v2
+	if !r.flagSetOwner.IsSet() {
 		r.eei.AddReturnMessage("invalid method to call")
 		return vmcommon.UserError
 	}
@@ -1491,11 +1347,11 @@ func (r *stakingSC) setOwner(args *vmcommon.ContractCallInput) vmcommon.ReturnCo
 
 // EpochConfirmed is called whenever a new epoch is confirmed
 func (r *stakingSC) EpochConfirmed(epoch uint32) {
-	r.flagStake.Toggle(epoch >= r.enableStakingEpoch)
-	log.Debug("stakingSC: stake/unstake/unbond", "enabled", r.flagStake.IsSet())
+	r.flagEnableStaking.Toggle(epoch >= r.enableStakingEpoch)
+	log.Debug("stakingSC: stake/unstake/unbond", "enabled", r.flagEnableStaking.IsSet())
 
-	r.flagStakingV2.Toggle(epoch >= r.stakingV2Epoch)
-	log.Debug("stakingSC: staking v2", "enabled", r.flagStakingV2.IsSet())
+	r.flagSetOwner.Toggle(epoch >= r.stakingV2Epoch)
+	log.Debug("stakingSC: set owner", "enabled", r.flagSetOwner.IsSet())
 }
 
 // IsInterfaceNil verifies if the underlying object is nil or not
